fix:constantとtypesのディレクトリを外に出す
